feat: Update default headers in ProxyHandler
The code changes update the default headers in the `ProxyHandler` function in the `handler.go` file. The `Referer` header is changed to `"https://scansmangas.me"` when the `urlString` has a prefix of `"https://scansmangas.me"`. This change improves the handling of requests to the `scansmangas.me` domain.

Note: Please remove any meta information like issue references, tags, or author names from the commit message.